requirements.txt
fastapi
uvicorn
streamlit
requests
faiss-cpu
sentence-transformers
pypdf
langchain
langgraph
